feat(chap3): Migrate to clean archtecture (test)
